+++ examples/basics.rs
@@ -25,10 +25,10 @@ fn main() {
     grad_example();
     println!("Cuda available: {}", tch::Cuda::is_available());
     println!("Cudnn available: {}", tch::Cuda::cudnn_is_available());
-    println!("{}", t);
+    println!("{t}");
     let t = Tensor::randn(&[1000], kind::FLOAT_CPU);
-    println!("{}", t);
+    println!("{t}");
     let t = Tensor::randn(&[1000, 1000], kind::FLOAT_CPU);
-    println!("{}", t);
+    println!("{t}");
     println!("{}", t * 100)
 }