+++ examples/tensor-tools.rs
@@ -18,16 +18,21 @@ pub fn main() -> Result<()> {
             for filename in args.iter().skip(2) {
                 if filename.ends_with(".npy") {
                     let tensor = tch::Tensor::read_npy(filename)?;
-                    println!("{}: {:?}", filename, tensor);
+                    println!("{filename}: {tensor:?}");
                 } else if filename.ends_with(".npz") {
                     let tensors = tch::Tensor::read_npz(filename)?;
                     for (name, tensor) in tensors.iter() {
-                        println!("{}: {} {:?}", filename, name, tensor)
+                        println!("{filename}: {name} {tensor:?}")
                     }
                 } else if filename.ends_with(".ot") {
                     let tensors = tch::Tensor::load_multi(filename)?;
                     for (name, tensor) in tensors.iter() {
-                        println!("{}: {} {:?}", filename, name, tensor)
+                        println!("{filename}: {name} {tensor:?}")
+                    }
+                } else if filename.ends_with(".bin") || filename.ends_with(".zip") {
+                    let tensors = tch::Tensor::loadz_multi(filename)?;
+                    for (name, tensor) in tensors.iter() {
+                        println!("{filename}: {name} {tensor:?}")
                     }
                 } else {
                     bail!("unhandled file {}", filename);
@@ -42,11 +47,13 @@ pub fn main() -> Result<()> {
                 tch::Tensor::read_npz(src_filename)?
             } else if src_filename.ends_with(".ot") {
                 tch::Tensor::load_multi(src_filename)?
+            } else if src_filename.ends_with(".bin") || src_filename.ends_with(".zip") {
+                tch::Tensor::loadz_multi(src_filename)?
             } else {
                 bail!("unhandled file {}", src_filename)
             };
             for (name, tensor) in tensors.iter() {
-                println!("{}: {} {:?}", src_filename, name, tensor)
+                println!("{src_filename}: {name} {tensor:?}")
             }
             if dst_filename.ends_with(".npz") {
                 tch::Tensor::write_npz(&tensors, dst_filename)?