+++ src/vision/image.rs
@@ -40,7 +40,7 @@ pub fn save<T: AsRef<Path>>(t: &Tensor, path: T) -> Result<(), TchError> {
     match t.size().as_slice() {
         [1, _, _, _] => save_hwc(&chw_to_hwc(&t.squeeze_dim(0)).to_device(Device::Cpu), path),
         [_, _, _] => save_hwc(&chw_to_hwc(&t).to_device(Device::Cpu), path),
-        sz => Err(TchError::FileFormat(format!("unexpected size for image tensor {:?}", sz))),
+        sz => Err(TchError::FileFormat(format!("unexpected size for image tensor {sz:?}"))),
     }
 }