+++ tests/var_store.rs
@@ -233,13 +233,13 @@ fn init_test() {
     let uniform = vs.root().var("t5", &[100], Init::Uniform { lo: 1.0, up: 2.0 });
     let uniform_min = f64::from(&uniform.min());
     let uniform_max = f64::from(&uniform.max());
-    assert!(uniform_min >= 1., "min {}", uniform_min);
-    assert!(uniform_max <= 2., "max {}", uniform_max);
+    assert!(uniform_min >= 1., "{}", "min {uniform_min}");
+    assert!(uniform_max <= 2., "{}", "max {uniform_max}");
     let uniform_std = f64::from(&uniform.std(true));
-    assert!(uniform_std > 0.15 && uniform_std < 0.35, "std {}", uniform_std);
+    assert!(uniform_std > 0.15 && uniform_std < 0.35, "{}", "std {uniform_std}");
     let normal = vs.root().var("normal", &[100], Init::Randn { mean: 0., stdev: 0.02 });
     let normal_std = f64::from(&normal.std(true));
-    assert!(normal_std <= 0.03, "std {}", normal_std);
+    assert!(normal_std <= 0.03, "{}", "std {normal_std}");
     let mut vs2 = VarStore::new(Device::Cpu);
     let ones = vs2.root().ones("t1", &[3]);
     assert_eq!(Vec::<f64>::from(&ones), [1., 1., 1.]);
@@ -247,7 +247,11 @@ fn init_test() {
     assert_eq!(Vec::<f64>::from(&ones), [0., 0., 0.]);
     let ortho = vs.root().var("orthogonal", &[100, 100], Init::Orthogonal { gain: 2.0 });
     let ortho_norm = f64::from(ortho.linalg_norm_ord_str("fro", None, true, Kind::Float));
-    assert!(f64::abs(ortho_norm - 20.) < 1e-5, "ortho_norm initialization failed {}", ortho_norm);
+    assert!(
+        f64::abs(ortho_norm - 20.) < 1e-5,
+        "{}",
+        "ortho_norm initialization failed {ortho_norm}"
+    );
     let ortho_shape_fail = tch::nn::f_init(Init::Orthogonal { gain: 1.0 }, &[10], Device::Cpu);
     assert!(ortho_shape_fail.is_err());
     let kaiming_u = vs.root().var("kaiming_u", &[20, 100], nn::init::DEFAULT_KAIMING_UNIFORM);
@@ -259,44 +263,44 @@ fn init_test() {
     assert!(f64::abs(f64::from(kaiming_n.std(true)) - (0.02f64).sqrt()) < 3e-3);
 }
 
-fn check_param_group(mut opt: tch::nn::Optimizer, foo: Tensor, bar: Tensor) {
+fn check_param_group(mut opt: tch::nn::Optimizer, var_foo: Tensor, var_bar: Tensor) {
     opt.set_lr(0.1);
     opt.set_lr_group(0, 0.);
     for _idx in 1..100 {
-        let loss = (&foo + &bar).mse_loss(&Tensor::from(0.42f32), tch::Reduction::Mean);
+        let loss = (&var_foo + &var_bar).mse_loss(&Tensor::from(0.42f32), tch::Reduction::Mean);
         opt.backward_step(&loss);
     }
-    assert_eq!(format!("{:.2}", f64::from(&foo)), "0.00");
-    assert_eq!(format!("{:.2}", f64::from(&bar)), "0.42");
+    assert_eq!(format!("{:.2}", f64::from(&var_foo)), "0.00");
+    assert_eq!(format!("{:.2}", f64::from(&var_bar)), "0.42");
     opt.set_lr_group(0, 0.1);
     for _idx in 1..100 {
-        let loss = (&foo + &bar).mse_loss(&Tensor::from(0f32), tch::Reduction::Mean);
+        let loss = (&var_foo + &var_bar).mse_loss(&Tensor::from(0f32), tch::Reduction::Mean);
         opt.backward_step(&loss);
     }
-    assert_eq!(format!("{:.2}", f64::from(&foo)), "-0.21");
-    assert_eq!(format!("{:.2}", f64::from(&bar)), "0.21");
+    assert_eq!(format!("{:.2}", f64::from(&var_foo)), "-0.21");
+    assert_eq!(format!("{:.2}", f64::from(&var_bar)), "0.21");
     opt.set_lr_group(7, 0.);
     for _idx in 1..100 {
-        let loss = (&foo + &bar).mse_loss(&Tensor::from(0.22f32), tch::Reduction::Mean);
+        let loss = (&var_foo + &var_bar).mse_loss(&Tensor::from(0.22f32), tch::Reduction::Mean);
         opt.backward_step(&loss);
     }
-    assert_eq!(format!("{:.2}", f64::from(&foo)), "0.01");
-    assert_eq!(format!("{:.2}", f64::from(&bar)), "0.21");
+    assert_eq!(format!("{:.2}", f64::from(&var_foo)), "0.01");
+    assert_eq!(format!("{:.2}", f64::from(&var_bar)), "0.21");
     // The following sets the learning rate for both groups.
     opt.set_lr(0.);
     for _idx in 1..100 {
-        let loss = (&foo + &bar).mse_loss(&Tensor::from(0.42f32), tch::Reduction::Mean);
+        let loss = (&var_foo + &var_bar).mse_loss(&Tensor::from(0.42f32), tch::Reduction::Mean);
         opt.backward_step(&loss);
     }
-    assert_eq!(format!("{:.2}", f64::from(&foo)), "0.01");
-    assert_eq!(format!("{:.2}", f64::from(&bar)), "0.21");
+    assert_eq!(format!("{:.2}", f64::from(&var_foo)), "0.01");
+    assert_eq!(format!("{:.2}", f64::from(&var_bar)), "0.21");
     opt.set_lr(0.1);
     for _idx in 1..100 {
-        let loss = (&foo + &bar).mse_loss(&Tensor::from(0.42f32), tch::Reduction::Mean);
+        let loss = (&var_foo + &var_bar).mse_loss(&Tensor::from(0.42f32), tch::Reduction::Mean);
         opt.backward_step(&loss);
     }
-    assert_eq!(format!("{:.2}", f64::from(&foo)), "0.11");
-    assert_eq!(format!("{:.2}", f64::from(&bar)), "0.31");
+    assert_eq!(format!("{:.2}", f64::from(&var_foo)), "0.11");
+    assert_eq!(format!("{:.2}", f64::from(&var_bar)), "0.31");
 }
 
 #[test]
@@ -305,9 +309,9 @@ fn param_group() {
     let vs = VarStore::new(Device::Cpu);
     let opt = tch::nn::Sgd::default().build(&vs, 1.0).unwrap();
     let root = vs.root();
-    let foo = root.set_group(0).zeros("foo", &[]);
-    let bar = root.set_group(7).zeros("bar", &[]);
-    check_param_group(opt, foo, bar);
+    let var_foo = root.set_group(0).zeros("var_foo", &[]);
+    let var_bar = root.set_group(7).zeros("var_bar", &[]);
+    check_param_group(opt, var_foo, var_bar);
 }
 
 #[test]
@@ -347,28 +351,28 @@ fn param_group_weight_decay() {
     let mut opt = tch::nn::Sgd::default().build(&vs, 0.0).unwrap();
     opt.set_lr(0.1);
     let root = vs.root();
-    let foo = root.set_group(0).zeros("foo", &[]);
-    let bar = root.set_group(7).zeros("bar", &[]);
+    let var_foo = root.set_group(0).zeros("var_foo", &[]);
+    let var_bar = root.set_group(7).zeros("var_bar", &[]);
     for _idx in 1..100 {
-        let loss = (&foo + &bar).mse_loss(&Tensor::from(1f32), tch::Reduction::Mean);
+        let loss = (&var_foo + &var_bar).mse_loss(&Tensor::from(1f32), tch::Reduction::Mean);
         opt.backward_step(&loss);
     }
-    assert_eq!(format!("{:.2}", f64::from(&foo)), "0.50");
-    assert_eq!(format!("{:.2}", f64::from(&bar)), "0.50");
+    assert_eq!(format!("{:.2}", f64::from(&var_foo)), "0.50");
+    assert_eq!(format!("{:.2}", f64::from(&var_bar)), "0.50");
     opt.set_weight_decay(0.1);
     for _idx in 1..100 {
-        let loss = (&foo + &bar).mse_loss(&Tensor::from(1f32), tch::Reduction::Mean);
+        let loss = (&var_foo + &var_bar).mse_loss(&Tensor::from(1f32), tch::Reduction::Mean);
         opt.backward_step(&loss);
     }
-    assert_eq!(format!("{:.2}", f64::from(&foo)), "0.49");
-    assert_eq!(format!("{:.2}", f64::from(&bar)), "0.49");
+    assert_eq!(format!("{:.2}", f64::from(&var_foo)), "0.49");
+    assert_eq!(format!("{:.2}", f64::from(&var_bar)), "0.49");
     opt.set_weight_decay_group(7, 0.);
     for _idx in 1..100 {
-        let loss = (&foo + &bar).mse_loss(&Tensor::from(1f32), tch::Reduction::Mean);
+        let loss = (&var_foo + &var_bar).mse_loss(&Tensor::from(1f32), tch::Reduction::Mean);
         opt.backward_step(&loss);
     }
-    assert_eq!(format!("{:.2}", f64::from(&foo)), "0.30");
-    assert_eq!(format!("{:.2}", f64::from(&bar)), "0.69");
+    assert_eq!(format!("{:.2}", f64::from(&var_foo)), "0.30");
+    assert_eq!(format!("{:.2}", f64::from(&var_bar)), "0.69");
 }
 
 #[test]