+++ torch-sys/libtch/torch_api.h
@@ -84,6 +84,8 @@ void at_load_multi(tensor *tensors, char **tensor_names, int ntensors, char *fil
 /* [at_load_multi_] takes as input an array of allocation [tensors]. */
 void at_load_multi_(tensor *tensors, char **tensor_names, int ntensors, char *filename);
 
+void at_loadz_callback(char *filename, void *data, void (*f)(void *, char *, tensor));
+void at_loadz_callback_with_device(char *filename, void *data, void (*f)(void *, char *, tensor), int device_id);
 void at_load_callback(char *filename, void *data, void (*f)(void *, char *, tensor));
 void at_load_callback_with_device(char *filename, void *data, void (*f)(void *, char *, tensor), int device_id);
 void at_load_from_stream_callback(void *stream_ptr, void *data, void (*f)(void *, char *, tensor), bool enable_device_id, int device_id);
@@ -111,11 +113,15 @@ void at_run_backward(tensor *tensors,
 optimizer ato_adam(double learning_rate,
                    double beta1,
                    double beta2,
-                   double weight_decay);
+                   double weight_decay,
+                   double eps,
+                   bool amsgrad);
 optimizer ato_adamw(double learning_rate,
                    double beta1,
                    double beta2,
-                   double weight_decay);
+                   double weight_decay,
+                   double eps,
+                   bool amsgrad);
 optimizer ato_rms_prop(double learning_rate,
                        double alpha,
                        double eps,